+++ packages/vue/src/__tests__/FormKit.spec.ts
@@ -3,6 +3,7 @@ import { flushPromises, mount } from '@vue/test-utils'
 import FormKit from '../FormKit'
 import { plugin } from '../plugin'
 import defaultConfig from '../defaultConfig'
+import { FormKitNode } from '@formkit/core'
 
 describe('props', () => {
   it('can display prop-defined errors', async () => {
@@ -61,5 +62,66 @@ describe('v-model', () => {
     wrapper.find('input').setValue('jon')
     await flushPromises()
     expect(wrapper.vm.$data.name).toBe('jon')
+  })
+})
+
+describe('events', () => {
+  it('emits the node as soon as it is created', () => {
+    const wrapper = mount(
+      {
+        template: '<FormKit @node="e => { node = e }" />',
+        data() {
+          return {
+            node: null as null | FormKitNode<any>,
+          }
+        },
+      },
+      {
+        global: {
+          plugins: [[plugin, defaultConfig]],
+        },
+      }
+    )
+    expect(wrapper.vm.node?.__FKNode__).toBe(true)
+  })
+})
+
+describe('validation', () => {
+  it('shows validation errors on a standard input', () => {
+    const wrapper = mount(FormKit, {
+      props: {
+        validation: 'required|length:5',
+      },
+      global: {
+        plugins: [[plugin, defaultConfig]],
+      },
+    })
+    expect(wrapper.html()).toContain('<li>')
+  })
+
+  it('can use arbitrarily created validation rules and messages', () => {
+    const wrapper = mount(
+      {
+        template: `
+          <FormKit
+            label="ABC"
+            validation="abc"
+            :validation-rules="{
+              abc: ({ value }) => value === 'abc'
+            }"
+            :validation-messages="{
+              abc: ({ name }) => name + ' should be abc'
+            }"
+            value="foo"
+          />
+        `,
+      },
+      {
+        global: {
+          plugins: [[plugin, defaultConfig]],
+        },
+      }
+    )
+    expect(wrapper.html()).toContain('<li>ABC should be abc</li>')
   })
 })